only 5 failing tests, score of 3 achieved.

+++ RalesEngine/db/migrate/20160222222659_create_invoice_items.rb
@@ -4,7 +4,7 @@ class CreateInvoiceItems < ActiveRecord::Migration
       t.integer :item_id
       t.integer :invoice_id
       t.integer :quantity
-      t.integer :unit_price
+      t.decimal :unit_price
       t.datetime :created_at
       t.datetime :updated_at
     end
